Move middleware tests to a separate namespace
